requirements.txt
ansicolors==1.1.8
beautifulsoup4==4.9.1
boto3==1.5.32
botocore==1.8.50
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-log==0.3.2
contextlib2==0.6.0.post1
docutils==0.16
idna==2.10
jmespath==0.9.3
policy-sentry==0.8.2
python-dateutil==2.8.1
PyYAML==4.2b4
requests==2.24.0
s3transfer==0.1.13
schema==0.7.2
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
